RBAC work
Accessors have been moved to a seperate folder to make the code more managable.

I've refactored a few things with roles, i think it's kind of silly to have object specific permissions for entire roles in this context, so i've setup the role schema to allow no object permission to be specified. I'll leave this field in there for fine-grained user account specific access control (which i haven't implemented yet).

I've also added the role IDs to the data stored in the redux container.