front end authorization
i've setup some basic authorization which prevents non-admin users from seeing the item creation options on the navbar, and prevents users from seeing the content of these pages if they aren't the appropriate rank. I've done this through the use of redux (to hold the role information on the logged in user) and an AuthorizedRoute component (that simply prevents users who aren't allowed roles from viewing the page).