init
One of the pros of using a non-relational database is the fact that you can optimize it for queries. For example, instead of needing to join a bunch of tables to pull the data you want, you can simple store the data within the documents you are querying so that it gets pulled out with the document.

With this in mind, this increment seeks to optimize media database queries; the back-end shouldn't need to "package media" with the items every time a request is made. This means that the "media" api methods have been removed, as they are no-longer needed.