cart work
cart now has a button to proceed to checkout. items in checkout are pulled from the redux cache automatically (this may be changed to a server request if we decide to do server-side carts). The items still need to be sent to the orders server before the payment process is complete.

There's also a lot of cleanup needed.